third_party/WebKit/LayoutTests/plugins/npruntime/browser-object-identity-expected.txt

Test that plugin doesn't get a new browser object instance each time

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS embed.getAndForgetRememberedObject().foo is 'bar'
PASS embed.refCount(obj) is 1
PASS embed.refCount(obj) is 1
PASS embed.refCount(obj) is 2
PASS embed.getRememberedObject() is obj
PASS embed.getRememberedObject() is obj
FAIL embed.refCount(obj) should be 2. Was 4.
PASS embed.getAndForgetRememberedObject() is obj
FAIL embed.refCount(obj) should be 1. Was 4.
PASS successfullyParsed is true

TEST COMPLETE
